refactor(member): 기존 redis 코드 리팩토링
