+++ member-service/src/main/java/com/example/memberservice/auth/email/service/EmailAuthService.java
@@ -1,0 +1,18 @@
+package com.example.memberservice.auth.email.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class EmailAuthService {
+
+    //이메일 전송 객체 만들기
+    
+    
+    //이메일 전송
+    
+    
+    //이메일 확인
+    
+    
+    
+}
