+++ member-service/src/main/java/com/example/memberservice/common/security/handler/OAuthLoginSuccessHandler.java
@@ -1,8 +1,9 @@
 package com.example.memberservice.common.security.handler;
 
 
+import com.example.memberservice.auth.token.repository.RefreshTokenRedisRepository;
 import com.example.memberservice.common.exception.BusinessException;
-import com.example.memberservice.common.redis.service.RedisSingleDataService;
+import com.example.memberservice.common.redis.service.RedisSingleDataRepository;
 import com.example.memberservice.common.security.jwt.JwtProperties;
 import com.example.memberservice.common.security.jwt.JwtTokenGenerator;
 import com.example.memberservice.common.security.model.dto.CustomOAuth2UserDto;
@@ -27,7 +28,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 public class OAuthLoginSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
 
-    private final RedisSingleDataService redisSingleDataService;
+    private final RefreshTokenRedisRepository refreshTokenRedisRepository;
 
     private final JwtTokenGenerator tokenGenerator;
 
@@ -59,7 +60,7 @@ public class OAuthLoginSuccessHandler extends SimpleUrlAuthenticationSuccessHand
 
         log.info("TOKEN:%s".formatted(memberCode));
         try {
-            redisSingleDataService.setSingleData(memberCode, refreshToken, jwtProperties.getRefreshTokenTtl());
+            refreshTokenRedisRepository.setSingleData(memberCode, refreshToken, jwtProperties.getRefreshTokenTtl());
 
         } catch (BusinessException e) {
             oAuthLoginFailureHandler.onAuthenticationFailure(request, response,
