+++ member-service/src/test/java/com/example/memberservice/common/redis/service/RefreshTokenRedisServiceTest.java
@@ -2,9 +2,8 @@ package com.example.memberservice.common.redis.service;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.netflix.discovery.converters.Auto;
+import com.example.memberservice.auth.token.repository.RefreshTokenRedisRepository;
 import java.util.Optional;
-import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +13,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 class RefreshTokenRedisServiceTest {
 
     @Autowired
-    private RedisSingleDataService redisSingleDataService;
+    private RefreshTokenRedisRepository refreshTokenRedisRepository;
 
 
     @Test
@@ -26,16 +25,16 @@ class RefreshTokenRedisServiceTest {
         String value = "value123";
 
         //When
-        redisSingleDataService.setSingleData(key, value,50000);
+        refreshTokenRedisRepository.setSingleData(key, value,50000);
 
         //Then
-        Optional<String> saveData = redisSingleDataService.getSingleData(key);
+        Optional<String> saveData = refreshTokenRedisRepository.getSingleData(key);
 
         assertEquals("value123", saveData.orElse("FAIL"));
 
-        redisSingleDataService.deleteSingleData(key);
+        refreshTokenRedisRepository.deleteSingleData(key);
 
-        Optional<String> deletedData = redisSingleDataService.getSingleData(key);
+        Optional<String> deletedData = refreshTokenRedisRepository.getSingleData(key);
 
         assertEquals("FAIL", deletedData.orElse("FAIL"));
     }
